@@ -7,17 +7,21 @@
 
 import Foundation
 
-public class DownloadTaskURLSessionDelegate: NSObject, URLSessionTaskDelegate {
+public final class DownloadTaskURLSessionDelegate: NSObject, URLSessionTaskDelegate, @unchecked Sendable {
     
     var progressObserver: NSKeyValueObservation?
     
-    var updateProgress: (Progress, NSKeyValueObservedChange<Double>) -> ()
+    let updateProgress: @Sendable (Progress, NSKeyValueObservedChange<Double>) -> ()
     
-    init(_ updateProgress: @escaping (Progress, NSKeyValueObservedChange<Double>) -> ()) {
+    private let queue = DispatchQueue(label: "downloadTaskDelegateLockQueue-" + UUID().uuidString)
+    
+    init(_ updateProgress: @Sendable @escaping (Progress, NSKeyValueObservedChange<Double>) -> ()) {
         self.updateProgress = updateProgress
     }
     
     public func urlSession(_ session: URLSession, didCreateTask task: URLSessionTask) {
-        progressObserver = task.progress.observe(\.fractionCompleted, changeHandler: updateProgress)
+        self.queue.sync {
+            self.progressObserver = task.progress.observe(\.fractionCompleted, changeHandler: updateProgress)
+        }
     }
 }
