@@ -5,7 +5,7 @@
 //  Created by Leo Wehrfritz on 20.01.25.
 //
 
-public enum AdfOption: String {
+public enum AdfOption: String, Sendable {
     case detectPaperLoaded = "DetectPaperLoaded"
     case selectSinglePage = "SelectSinglePage"
     case duplex = "Duplex"
