@@ -5,7 +5,7 @@
 //  Created by Leo Wehrfritz on 20.01.25.
 //
 
-public enum BinaryRendering: String {
+public enum BinaryRendering: String, Sendable {
     case threshold = "Threshold"
     case halftone = "Halftone"
 }
