@@ -5,7 +5,7 @@
 //  Created by Leo Wehrfritz on 20.01.25.
 //
 
-public enum ColorMode: String {
+public enum ColorMode: String, Sendable {
     case blackAndWhite = "BlackAndWhite1"
     case grayscale8 = "Grayscale8"
     case grayscale16 = "Grayscale16"
