@@ -5,7 +5,7 @@
 //  Created by Leo Wehrfritz on 23.01.25.
 //
 
-enum XMLDecodingError: Error {
+enum XMLDecodingError: Error, Sendable {
     case couldntInstantiateObject
     case missingKey(String)
     case unexptedType(Any.Type, String)
